refactored habit components
