+++ components/index.ts
@@ -1,12 +1,12 @@
 import React from "react";
 export { default as Container } from "./Container";
 export { default as DaysHeading } from "./DaysHeading";
-export { default as DisplayHabits } from "./DisplayHabits";
 export { default as FormContainer } from "./FormContainer";
 export { default as FormGroup } from "./FormGroup";
 export { default as FormRow } from "./FormRow";
-export { default as HabitDailyCell } from "./HabitDailyCell";
-export { default as HabitDailyRow } from "./HabitDailyRow";
+export { default as DailyCell } from "./habits/DailyCell";
+export { default as DailyRow } from "./habits/DailyRow";
+export { default as DisplayHabits } from "./habits/Display";
 export { default as Heading1 } from "./Heading1";
 export { default as HeadingMobile } from "./HeadingMobile";
 export { default as Layout } from "./Layout";
